src/tasks/tasks.service.spec.ts
import { Test } from '@nestjs/testing';
import { TasksRepository } from './tasks.repository';
import { TasksService } from './tasks.service';
import { NotFoundException } from '@nestjs/common';

const mockTasksRepository = () => ({
  getTasks: jest.fn(),
  findOneBy: jest.fn(),
});

const mockUser = {
  username: 'vinay',
  id: 'id123',
  password: 'password',
  tasks: [],
};

describe('TasksService', () => {
  let tasksService: TasksService;
  let tasksRepository;

  beforeEach(async () => {
    const module = await Test.createTestingModule({
      providers: [
        TasksService,
        { provide: TasksRepository, useFactory: mockTasksRepository },
      ],
    }).compile();

    tasksService = module.get(TasksService);
    tasksRepository = module.get(TasksRepository);
  });

  describe('getTasks', () => {
    it('calls TasksRepository.getTasks and returns the result', async () => {
      tasksRepository.getTasks.mockResolvedValue('somevalue');
      const result = await tasksService.getTasks(null, mockUser);
      expect(tasksRepository.getTasks).toHaveBeenCalled();
      expect(result).toEqual('somevalue');
    });
  });

  describe('getTasksById', () => {
    it('should call TasksRepository.getTasksById and return the result', async () => {
      tasksRepository.findOneBy.mockResolvedValue('someValue');
      const result = await tasksService.getTaskById('123', mockUser);
      expect(tasksRepository.findOneBy).toHaveBeenCalled();
      expect(result).toEqual('someValue');
    });

    it('should throw error if task with id not found', async () => {
      tasksRepository.findOneBy.mockResolvedValue(null);
      expect(tasksService.getTaskById('123', mockUser)).rejects.toThrow(
        NotFoundException,
      );
    });
  });
});
